+++ R/docx_cursor.R
@@ -25,21 +25,13 @@
 #' doc <- cursor_begin(doc)
 #' doc <- body_remove(doc)
 #'
-#' # Let add text at the beginning of the
-#' # paragraph containing text "paragraph 4"
 #' doc <- cursor_reach(doc, keyword = "paragraph 4")
-#' doc <- slip_in_text(doc, "This is ", pos = "before", style = "Default Paragraph Font")
 #'
-#' # move the cursor forward and end a section
+#' # move the cursor forward
 #' doc <- cursor_forward(doc)
-#' doc <- body_add_par(doc, "The section stop here", style = "Normal")
-#' doc <- body_end_section_landscape(doc)
 #'
 #' # move the cursor at the end of the document
 #' doc <- cursor_end(doc)
-#' doc <- body_add_par(doc, "The document ends now", style = "Normal")
-#'
-#' print(doc, target = tempfile(fileext = ".docx"))
 cursor_begin <- function( x ){
   x$doc_obj$cursor_begin()
   x